flask/requirements.txt
flask
flask_wtf
mysql-connector
bcrypt
uwsgi